requirements.txt
pywebview
pygobject
darkdetect